refactor(text-image-morph): extract a reusable text->image morph engine
The morph logic was scoped to "the little prince". Split it into a generic,
image-agnostic engine and make the prince a thin instance of it:

- new src/animations/text-image-morph/: the engine — TextImageMorph component +
  useTextImageMorph hook + sampling/layout. Takes any image + paragraph; atlas
  geometry is derived from the paragraph (no longer module-static). Bundled
  Newsreader serif lives here.
- src/animations/the-little-prince/ is now just the instance: prince.png,
  text.ts (the passage), and a screen that renders <TextImageMorph image=...
  paragraph=... />.

Perf: the page text no longer waits on the expensive picture sampling. Layout
is computed synchronously (text paints immediately); the readPixels + Poisson
sampling + prune is deferred to a post-paint macrotask, then the morph buffers
rebuild via a targetsReady bump. Fixes the slow initial text load.

Animation, formatting, and illustration quality are unchanged.
