+++ src/animations/text-image-morph/constants.ts
@@ -1,33 +1,6 @@
-// The Little Prince (Antoine de Saint-Exupéry) — the fox's farewell.
-// Themed on seeing with the heart, which ties to the picture reveal.
-// Paragraph breaks are marked with '\n' — the layout lays each paragraph on
-// its own line with a first-line indent (real book formatting).
-export const PARAGRAPH =
-  '“You are beautiful, but you are empty,” he went on. “One could not die ' +
-  'for you. To be sure, an ordinary passerby would think that my rose ' +
-  'looked just like you—the rose that belongs to me. But in herself alone ' +
-  'she is more important than all the hundreds of you other roses: because ' +
-  'it is she that I have watered; because it is she that I have put under ' +
-  'the glass globe; because it is she that I have sheltered behind the ' +
-  'screen; because it is for her that I have killed the caterpillars ' +
-  '(except the two or three that we saved to become butterflies); because ' +
-  'it is she that I have listened to, when she grumbled, or boasted, or ' +
-  'even sometimes when she said nothing. Because she is my rose.”\n' +
-  'And he went back to meet the fox.\n' +
-  '“Goodbye,” he said.\n' +
-  '“Goodbye,” said the fox.\n' +
-  '“And now here is my secret, a very simple secret: It is only with the ' +
-  'heart that one can see rightly; what is essential is invisible to the ' +
-  'eye.”\n' +
-  'What is essential is invisible to the eye, the little prince repeated, ' +
-  'so that hember.\n' +
-  '“It is the time you have wasted for your rose that makes your rose so ' +
-  'important.”\n' +
-  '“It is the time I have wasted for my rose…” the little prince repeated, ' +
-  'so that he would be sure to remember.\n' +
-  '“Men have forgotten this truth,” said the fox. “But you must not forget ' +
-  'it. You become responsible, forever, for what you have tamed. You are ' +
-  'responsible for your rose…”';
+// Text -> image morph: a page of text whose letters fly together to form a
+// picture sampled from an arbitrary image. These are the generic look/feel and
+// sampling knobs; the actual text + image are supplied per instance.
 
 // Page text styling
 export const PAGE_BG = '#f7f4ec'; // soft warm white
@@ -38,7 +11,7 @@ export const GLYPH_CELL = 40;
 export const GLYPH_FONT_SIZE = 30;
 export const GLYPH_PAD = 3; // left bearing of the glyph inside its cell
 
-// Display scale of a glyph on the page (book text size)
+// Display scale of a glyph on the page vs. in the assembled picture
 export const PAGE_GLYPH_SCALE = 0.48;
 export const PICTURE_GLYPH_SCALE = 0.44;
 
@@ -55,8 +28,8 @@ export const LUM_HI = 0.22;
 
 // 'ink' params: ink = max(1 - lum, saturation * SAT_WEIGHT).
 export const INK_FLOOR = 0.18; // below this -> background, no letters
-// Lower weight => bright saturated bits (yellow stars) don't dominate; the
-// dark-outlined main subject (prince) stays the densest focal mass.
+// Lower weight => bright saturated bits don't dominate; the dark-outlined main
+// subject stays the densest focal mass.
 export const SAT_WEIGHT = 0.45;
 
 export const SAMPLE_GAMMA = 1.4; // contrast: dense subject, thin background
