+++ src/animations/text-image-morph/index.ts
@@ -1,0 +1,7 @@
+export { TextImageMorph } from './text-image-morph';
+export { useTextImageMorph } from './use-text-image-morph';
+export type {
+  Particle,
+  Atlas,
+  TextImageMorphData,
+} from './use-text-image-morph';
