+++ src/animations/the-little-prince/text.ts
@@ -1,0 +1,30 @@
+// The Little Prince (Antoine de Saint-Exupéry) — the rose & the fox's farewell.
+// Themed on seeing with the heart, which ties to the picture reveal.
+// Paragraph breaks are marked with '\n' — the layout lays each paragraph on
+// its own line with a first-line indent (real book formatting).
+export const LITTLE_PRINCE_TEXT =
+  '“You are beautiful, but you are empty,” he went on. “One could not die ' +
+  'for you. To be sure, an ordinary passerby would think that my rose ' +
+  'looked just like you—the rose that belongs to me. But in herself alone ' +
+  'she is more important than all the hundreds of you other roses: because ' +
+  'it is she that I have watered; because it is she that I have put under ' +
+  'the glass globe; because it is she that I have sheltered behind the ' +
+  'screen; because it is for her that I have killed the caterpillars ' +
+  '(except the two or three that we saved to become butterflies); because ' +
+  'it is she that I have listened to, when she grumbled, or boasted, or ' +
+  'even sometimes when she said nothing. Because she is my rose.”\n' +
+  'And he went back to meet the fox.\n' +
+  '“Goodbye,” he said.\n' +
+  '“Goodbye,” said the fox.\n' +
+  '“And now here is my secret, a very simple secret: It is only with the ' +
+  'heart that one can see rightly; what is essential is invisible to the ' +
+  'eye.”\n' +
+  'What is essential is invisible to the eye, the little prince repeated, ' +
+  'so that hember.\n' +
+  '“It is the time you have wasted for your rose that makes your rose so ' +
+  'important.”\n' +
+  '“It is the time I have wasted for my rose…” the little prince repeated, ' +
+  'so that he would be sure to remember.\n' +
+  '“Men have forgotten this truth,” said the fox. “But you must not forget ' +
+  'it. You become responsible, forever, for what you have tamed. You are ' +
+  'responsible for your rose…”';
